requirements.txt
aiosqlite==0.17.0
annotated-types==0.7.0
asttokens==2.4.1
async-timeout==4.0.3
asyncpg==0.29.0
click==8.1.7
decorator==5.1.1
dictdiffer==0.9.0
exceptiongroup==1.2.2
executing==2.0.1
ipython==8.26.0
iso8601==1.1.0
jedi==0.19.1
matplotlib-inline==0.1.7
parso==0.8.4
pexpect==4.9.0
prompt_toolkit==3.0.47
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
pypika-tortoise==0.1.6
pytz==2024.1
six==1.16.0
stack-data==0.6.3
toml==0.10.2
tomlkit==0.13.2
tortoise-orm==0.21.6
traitlets==5.14.3
typing_extensions==4.12.2
wcwidth==0.2.13
